+++ src/redturtle/prenotazioni/utilities/dateutils.py
@@ -2,8 +2,6 @@
 from datetime import datetime
 from datetime import timedelta
 from redturtle.prenotazioni import tznow
-from plone.restapi.serializer.converters import datetimelike_to_iso
-from DateTime import DateTime
 
 
 def exceedes_date_limit(data, future_days):
@@ -23,16 +21,3 @@ def exceedes_date_limit(data, future_days):
     if booking_date <= date_limit:
         return False
     return True
-
-
-def datetimelike_to_iso_tz(value, tzinfo):
-    """se sul db non c'è timezone la data è stata salvata con il
-    default_timezone
-    """
-    if value is None:
-        return None
-    if isinstance(value, DateTime):
-        return datetimelike_to_iso(value)
-    if value.tzinfo is None:
-        return value.astimezone(tzinfo).isoformat()
-    return value.isoformat()